+++ Sources/RevenueCatKit/RevenueCatSDKAdapter.swift
@@ -111,6 +111,12 @@ final class RevenueCatSDKAdapter: RevenueCatProviding {
                     localizedPrice: package.storeProduct.localizedPriceString,
                     currencyCode: package.storeProduct.currencyCode,
                     subscriptionPeriod: package.storeProduct.subscriptionPeriod?.companyValue,
+                    introductoryOffer: package.storeProduct.introductoryDiscount.map {
+                        IntroductoryOffer(
+                            localizedPrice: $0.localizedPriceString,
+                            paymentMode: $0.paymentMode.companyValue
+                        )
+                    },
                     productID: package.storeProduct.productIdentifier
                 )
             }
@@ -254,6 +260,17 @@ final class RevenueCatSDKAdapter: RevenueCatProviding {
             return .operationInProgress
         default:
             return .unknown
+        }
+    }
+}
+
+private extension StoreProductDiscount.PaymentMode {
+    var companyValue: IntroductoryOffer.PaymentMode {
+        switch self {
+        case .payAsYouGo: .payAsYouGo
+        case .payUpFront: .payUpFront
+        case .freeTrial: .freeTrial
+        @unknown default: .unknown
         }
     }
 }